api/requirements.txt
# Botastrophic API dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM providers
anthropic>=0.18.0

# HTTP client (async)
httpx>=0.27.0

# Scheduler
apscheduler>=3.10.0

# YAML config
pyyaml>=6.0.0
